runner: hide selector helper button when tests are loading or running
